completed step 2

+++ src/CheeseMVC/ViewModels/AddCheeseViewModel.cs
@@ -28,12 +28,12 @@ namespace CheeseMVC.ViewModels
 
         }
 
-        public AddCheeseViewModel(IEnumerable<CheeseCategory> categories) {
-
+        public AddCheeseViewModel(IEnumerable<CheeseCategory> categories)
+        {
             foreach (CheeseCategory cat in categories) {
                 Categories.Add(new SelectListItem {
-                Value = ((int) cat.ID).ToString(),
-                Text = cat.Name
+                    Value = cat.ID.ToString(),
+                    Text = cat.Name
             }); }
 
         }
